fix: address SonarQube code smells across 6 files
- AI.Mcp: remove unnecessary (float) cast on Temperature
- DPoP: rephrase comment to avoid false positive on "commented out code"
- ConnectTokenEndpoints: remove unused principalFactory param from
  HandleClientCredentials (static method, not injected)
- OpenIddictEndpointsModule: use expression body (IDE0022)
- Privacy.Regulations: change int? to int in DefaultResponseDeadlineTracker
  (SubjectAccessRequestDays is required int, null path unreachable)
- Privacy.Regulations.Tests: use expression body for single-line test

Won't fix (per CLAUDE.md):
- brain-overload (3 methods): cognitive complexity in Observability,
  ConnectUserInfoEndpoints, OpenIddictServer — structural, not simplifiable
- cancellationToken unused in AccountLoginEndpoints: ASP.NET convention,
  handlers should keep CT for future async calls
- "Adequacy"/"Consent" literals: domain registration data, not code logic
- Skipped test: explicitly requested to keep

+++ src/Granit.AI.Mcp/Internal/McpSamplingChatClientAdapter.cs
@@ -95,7 +95,7 @@ internal sealed partial class McpSamplingChatClientAdapter(
 
         if (request.Temperature.HasValue)
         {
-            chatOptions.Temperature = (float)request.Temperature.Value;
+            chatOptions.Temperature = request.Temperature.Value;
         }
 
         if (request.ModelPreferences?.Hints is { Count: > 0 } hints)
